Guard playground localStorage access, extract ExamplePicker
Wraps all localStorage reads/writes in try/catch — sandboxed iframes
and strict privacy modes can throw SecurityError on access, which
would otherwise crash the useState initializer and prevent the
playground from mounting.

Also moves ExamplePicker into its own module to keep playground.tsx
under its line budget after the safe-storage helpers were added.
