+++ renderer/src/playground/playground.tsx
@@ -10,7 +10,6 @@ import { Toaster } from "sonner";
 import {
   Moon,
   Sun,
-  Search,
   Braces,
   Link,
   Check,
@@ -23,7 +22,6 @@ import { Button } from "@/ui/button";
 import { Badge } from "@/ui/badge";
 import { Alert, AlertDescription } from "@/ui/alert";
 import { Separator } from "@/ui/separator";
-import { Popover, PopoverContent, PopoverTrigger } from "@/ui/popover";
 import {
   Dialog,
   DialogContent,
@@ -33,8 +31,9 @@ import {
   DialogTitle,
 } from "@/ui/dialog";
 import { Editor } from "./editor";
+import { ExamplePicker } from "./example-picker";
 import { executePython, loadPyodideRuntime } from "./pyodide";
-import { EXAMPLES, type Example } from "./examples";
+import { type Example } from "./examples";
 import { ThemePicker } from "./theme-picker";
 import { buildThemeCss } from "../themes";
 import pako from "pako";
@@ -151,83 +150,22 @@ with Card():
 
 const DEBOUNCE_MS = 200;
 
-function ExamplePicker({ onSelect }: { onSelect: (ex: Example) => void }) {
-  const [open, setOpen] = useState(false);
-  const [filter, setFilter] = useState("");
-  const inputRef = useRef<HTMLInputElement>(null);
+/** Wrap localStorage reads — sandboxed iframes or strict privacy
+ *  modes can throw SecurityError on access. */
+function safeStorageGet(key: string): string | null {
+  try {
+    return localStorage.getItem(key);
+  } catch {
+    return null;
+  }
+}
 
-  const filtered = useMemo(() => {
-    if (!filter) return EXAMPLES;
-    const q = filter.toLowerCase();
-    return EXAMPLES.filter(
-      (ex) =>
-        ex.title.toLowerCase().includes(q) ||
-        ex.category.toLowerCase().includes(q),
-    );
-  }, [filter]);
-
-  const grouped = useMemo(() => {
-    const groups = new Map<string, Example[]>();
-    for (const ex of filtered) {
-      const list = groups.get(ex.category) ?? [];
-      list.push(ex);
-      groups.set(ex.category, list);
-    }
-    return groups;
-  }, [filtered]);
-
-  useEffect(() => {
-    if (open) {
-      setFilter("");
-      requestAnimationFrame(() => inputRef.current?.focus());
-    }
-  }, [open]);
-
-  return (
-    <Popover open={open} onOpenChange={setOpen}>
-      <PopoverTrigger className="inline-flex h-10 items-center gap-2 rounded-md border border-border bg-background px-3 text-sm text-muted-foreground hover:bg-accent hover:text-accent-foreground">
-        <Search className="h-4 w-4" />
-        Examples...
-      </PopoverTrigger>
-      <PopoverContent align="start" className="w-[320px] p-0">
-        <div className="border-b border-border px-3 py-2">
-          <input
-            ref={inputRef}
-            value={filter}
-            onChange={(e) => setFilter(e.target.value)}
-            placeholder="Filter examples..."
-            className="w-full bg-transparent text-sm outline-none placeholder:text-muted-foreground"
-          />
-        </div>
-        <div className="max-h-[300px] overflow-y-auto py-1">
-          {filtered.length === 0 && (
-            <div className="px-3 py-4 text-center text-sm text-muted-foreground">
-              No matching examples.
-            </div>
-          )}
-          {[...grouped.entries()].map(([category, items]) => (
-            <div key={category}>
-              <div className="px-3 py-1.5 text-xs font-medium text-muted-foreground">
-                {category}
-              </div>
-              {items.map((ex) => (
-                <button
-                  key={ex.title}
-                  className="flex w-full items-center px-3 py-1.5 text-left text-sm hover:bg-accent hover:text-accent-foreground"
-                  onClick={() => {
-                    onSelect(ex);
-                    setOpen(false);
-                  }}
-                >
-                  {ex.title}
-                </button>
-              ))}
-            </div>
-          ))}
-        </div>
-      </PopoverContent>
-    </Popover>
-  );
+function safeStorageSet(key: string, value: string): void {
+  try {
+    localStorage.setItem(key, value);
+  } catch {
+    // Ignore — persistence is best-effort.
+  }
 }
 
 export function Playground() {
@@ -248,10 +186,10 @@ export function Playground() {
     () => window.matchMedia("(prefers-color-scheme: dark)").matches,
   );
   const [editorCollapsed, setEditorCollapsed] = useState(
-    () => localStorage.getItem("pg-editor-collapsed") === "true",
+    () => safeStorageGet("pg-editor-collapsed") === "true",
   );
   const [editorWidth, setEditorWidth] = useState(() => {
-    const stored = parseFloat(localStorage.getItem("pg-editor-width") ?? "");
+    const stored = parseFloat(safeStorageGet("pg-editor-width") ?? "");
     return Number.isFinite(stored)
       ? Math.max(0.15, Math.min(0.85, stored))
       : 0.5;
@@ -278,10 +216,10 @@ export function Playground() {
 
   // Persist split layout preferences
   useEffect(() => {
-    localStorage.setItem("pg-editor-collapsed", String(editorCollapsed));
+    safeStorageSet("pg-editor-collapsed", String(editorCollapsed));
   }, [editorCollapsed]);
   useEffect(() => {
-    localStorage.setItem("pg-editor-width", editorWidth.toFixed(3));
+    safeStorageSet("pg-editor-width", editorWidth.toFixed(3));
   }, [editorWidth]);
 
   // Draggable split divider — uses pointer capture so move/up events
